Move Comment to components
