+++ app/containers/App/index.js
@@ -1,26 +1,7 @@
 import React, { Component } from 'react'
-import Remarkable from 'remarkable'
+import Comment from '../../components/Comment'
 import $ from 'jquery'
 import { connect } from 'react-redux'
-
-class Comment extends Component {
-  rawMarkup() {
-    const md = new Remarkable()
-    const rawMarkup = md.render(this.props.children.toString())
-    return { __html: rawMarkup }
-  }
-
-  render() {
-    return (
-      <div className="comment">
-        <h2 className="commentAuthor">
-          {this.props.author}
-        </h2>
-        <span dangerouslySetInnerHTML={this.rawMarkup()} />
-      </div>
-    )
-  }
-}
 
 class CommentList extends Component {
   render() {
